veri: fix bug in BVShr type inference (#95)
This PR fixes a bug in the Solver in dynamic width mode and onlywidths true,
where the y operand of a BVShr expression will not be visited. This can
cause type inference to fail. This problem came up when trying to use a
Sail-generated specification that contained a concat expression in the
right-hand-side operand.